+++ src/commands/log.py
@@ -3,6 +3,7 @@ import pathlib
 
 import typer
 
+from src.utils.cli import PROFILES_OPTION
 from src.utils.cli import PROJECTS_ARGUMENT
 from src.utils.cli import RUNNING_OPTION
 from src.utils.compose_rich import ComposeProject
@@ -22,6 +23,7 @@ def log_project(project: ComposeProject, options: Options, info: ProjectInfo):
     rich_run_compose(
         project.dir,
         project.file,
+        project.selected_profiles,
         command=[
             "logs",
             *(["-f"] if options.follow else []),
@@ -34,6 +36,7 @@ def log_project(project: ComposeProject, options: Options, info: ProjectInfo):
 
 def main(
     projects: list[pathlib.Path] = PROJECTS_ARGUMENT,
+    profiles: list[str] = PROFILES_OPTION,
     running: bool = RUNNING_OPTION,
     no_follow: bool = typer.Option(False, "--no-follow", "-q", help="Quit right after printing logs."),
 ):
@@ -43,6 +46,7 @@ def main(
 
     for project in get_compose_projects(
         projects,
+        profiles,
         ProjectSearchOptions(
             print_compose_errors=False,
             only_running=running,