+++ src/commands/up.py
@@ -3,6 +3,7 @@ import pathlib
 
 import typer
 
+from src.utils.cli import PROFILES_OPTION
 from src.utils.cli import PROJECTS_ARGUMENT
 from src.utils.cli import RUNNING_OPTION
 from src.utils.compose_rich import ComposeProject
@@ -27,6 +28,7 @@ def up_project(project: ComposeProject, options: Options, info: ProjectInfo):
         rich_run_compose(
             project.dir,
             project.file,
+            project.selected_profiles,
             command=[
                 "up",
                 *(["--remove-orphans"] if not options.no_remove_orphans else []),
@@ -42,6 +44,7 @@ def up_project(project: ComposeProject, options: Options, info: ProjectInfo):
         rich_run_compose(
             project.dir,
             project.file,
+            project.selected_profiles,
             command=["logs", "-f"],
             dry_run=options.dry_run,
             cmds=info.cmds,
@@ -51,8 +54,9 @@ def up_project(project: ComposeProject, options: Options, info: ProjectInfo):
 
 def main(  # noqa: CFQ002 (max arguments)
     projects: list[pathlib.Path] = PROJECTS_ARGUMENT,
+    profiles: list[str] = PROFILES_OPTION,
     running: bool = RUNNING_OPTION,
-    do_pull: bool = typer.Option(False, "--pull", "-p", help="Pull images before running."),
+    do_pull: bool = typer.Option(False, "--pull", help="Pull images before running."),
     do_log: bool = typer.Option(False, "--log", "-l", help="Also show logs."),
     no_build: bool = typer.Option(False, "--no-build", help="Don't build images before running."),
     no_remove_orphans: bool = typer.Option(False, "--no-remove-orphans", help="Keep orphans."),
@@ -66,6 +70,7 @@ def main(  # noqa: CFQ002 (max arguments)
 
     for project in get_compose_projects(
         projects,
+        profiles,
         ProjectSearchOptions(
             print_compose_errors=dry_run,
             only_running=running,